pyproject.toml
[project]
name = "django-databruce"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.1.7",
    "django-autocomplete-light>=3.12.1",
    "django-crispy-forms==2.3",
    "django-debug-toolbar==5.1.0",
    "django-filter==24.3",
    "django-select2>=8.4.0",
    "djangorestframework==3.15.2",
    "djangorestframework-datatables==0.7.2",
    "djlint>=1.36.4",
    "markdown==3.7",
    "psycopg[binary,pool]",
    "python-dotenv==1.1.0",
]
